feat: 서비스 인터페이스/구현체 분리 및 userId 필드 추가
