+++ build.gradle
@@ -30,11 +30,12 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
-    // Spring Boot BOM
-    runtimeOnly 'org.postgresql:postgresql'
+    // === Flyway (핵심 수정) ===
+    implementation 'org.flywaydb:flyway-core:11.7.2'
+    implementation 'org.flywaydb:flyway-database-postgresql:11.7.2'  // ✅ 필수
+    // ==========================
 
-    // Flyway
-    implementation 'org.flywaydb:flyway-core'
+    runtimeOnly 'org.postgresql:postgresql'                    // JDBC 드라이버
 
     implementation 'com.vladmihalcea:hibernate-types-60:2.21.1'
 
@@ -42,8 +43,8 @@ dependencies {
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
 
-    // API 문서 (springdoc-openapi)
-    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.2.0'
+    // API 문서
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.11'
 
     // Test
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
